Add bash code to fix a bug in the sssd.conf file on Debian Bullseye
On Debian Bullseye, the SSSD config file incorrectly contains a line
of the form "services = ..." in the sssd section.  This line is
unnecessary since we are using SystemD (as described in
https://manpages.debian.org/testing/sssd-common/sssd.conf.5.en.html#The_%5Bsssd%5D_section),
and on Debian Bullseye its presence causes a failure.  This is because
SystemD tries to enable and start those services, but they are only
meant to be triggered when the corresponding SystemD
*.socket units are activated.

Note that /usr/bin/lsb_release is a Debian-specific tool, so if it
doesn't exist then we know we are not on Debian.

The sssd.conf file is created during the FreeIPA install process, so
we cannot edit this file before this script is run.