Tweak the default krb5.conf file on Debian Bullseye
On Debian Bullseye, the default /etc/krb5.conf file uses a FILE value
for the default_ccache_name option.  The /krb5.conf file is then
overwritten when the 00_setup_freeipa.sh script is run, and the new
file uses a KEYRING value.  This causes a problem since before running
that script we kinit, which stores our keytab in a file that is then
ignored by Kerberos due to the new configuration.  To avoid this
wrinkle we simply adjust the default configuration to use the
preferred value before the script can ever be run.