Do not specify backports as the default_release option
This option does not work as I thought it did.  I thought it was
necessary in order to pull a package out of backports, but it actually
causes apt to _prefer_ backports over any other repo source.  In the
case of freeipa-client, the package is only available in backports so
it is unnecessary here, and using it causes all freeipa-client
dependencies to be installed from backports if at all possible.  This
leads to broken behavior, as I mentioned here:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1011070