Remove services line from sssd config on Ubuntu as well
Ubuntu has the same issue, so we should be removing this line there as
well.  Note that I also simplified the logic to perform the removal
regardless of release name, since the systemd-based services with
socket activation will have found its way into all supported releases
by now; it was introduced in version 1.x of sssd.